Example data for use with the phenoptr package

phenoptrExamples contains data sets and example code which demonstrate the use of the phenoptr package. phenoptr contains functions that make it easier to read and analyze data tables and images created by Akoya Biosciences' inForm software.

Details

phenoptrExamples contains selected inForm output from nine images taken from three samples of lung cancer tissue. To see the names of all of the files included for these samples, use the command list.files(system.file("extdata", "samples", package = "phenoptrExamples")).

For details on the file contents, see the package documentation online.

phenoptr and phenoptrExamples are part of the Akoya Biosciences Phenoptics family of Quantitative Pathology Research Solutions. For more information visit the Phenoptics home page.

See also